+++ packages/ruleset/src/functions/index.js
@@ -1,5 +1,6 @@
 const arrayOfArrays = require('./array-of-arrays');
 const checkMajorVersion = require('./check-major-version');
+const descriptionMentionsJSON = require('./description-mentions-json');
 const discriminator = require('./discriminator');
 const errorResponseSchema = require('./error-response-schema');
 const propertyCaseCollision = require('./property-case-collision');
@@ -13,6 +14,7 @@ const validTypeFormat = require('./valid-type-format');
 module.exports = {
   arrayOfArrays,
   checkMajorVersion,
+  descriptionMentionsJSON,
   discriminator,
   errorResponseSchema,
   propertyCaseCollision,